# docs/api.yml
openapi: 3.0.3
info:
  title: Docbox API
  version: 1.0.0
  description: API specification for docbox
paths:
...
  /box/{scope}/folder:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
    post:
      summary: Create a folder
      operationId: create_folder
      description: >
        Create a folder within a document box
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  description: Name of the folder to create
                folder_id:
                  $ref: "#/components/schemas/FolderId"
                  description: Parent folder ID (Folder to create within)
              required:
                - name
                - folder_id
      responses:
        "201":
          description: Folder created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: The provided parent folder was not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  /box/{scope}/folder/{folder_id}:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
      - $ref: "#/components/parameters/folder_id"
    get:
      summary: Get a folder
      operationId: get_folder
      description: >
        Request a specific folder, responds with the folder itself as well as the 
        children contained in the folder (Depth of 1, not recursive)
      responses:
        "200":
          description: Folder exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderResponse"
        "404":
          description: Folder not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update a folder
      operationId: update_folder
      description: >
        Update a folder, allows changing the folder name and parent folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  nullable: true
                  minLength: 1
                folder_id:
                  $ref: "#/components/schemas/FolderId"
                  nullable: true
      responses:
        "200":
          description: Folder updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Folder not found or target folder could not be found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a folder
      operationId: delete_folder
      description: >
        Deletes a folder and everything contained within the folder
      responses:
        "204":
          description: Folder deleted successfully
        "404":
          description: Folder not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /box/{scope}/link:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
    post:
      summary: Create a link
      operationId: create_link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                value:
                  type: string
                  minLength: 1
                folder_id:
                  $ref: "#/components/schemas/FolderId"
              required:
                - name
                - value
                - folder_id
      responses:
        "201":
          description: Link created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkWithExtra"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /box/{scope}/link/{link_id}:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
      - $ref: "#/components/parameters/link_id"
    get:
      summary: Get a link
      operationId: get_link
      responses:
        "200":
          description: Link retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkWithExtra"
        "404":
          description: Link not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update a link
      operationId: update_link
      description: >
        Update the properties of a specific link such as renaming, updating the link value,
        and moving the link to another folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  nullable: true
                  description: New name for the link if renaming the link
                value:
                  type: string
                  minLength: 1
                  nullable: true
                  description: New value for the link URL if changing the link
                folder_id:
                  allOf:
                    - $ref: "#/components/schemas/FolderId"
                    - nullable: true
                      description: The folder to move the link to if moving the link
      responses:
        "200":
          description: Link updated successfully
        "404":
          description: Link not found or destination folder not found when moving
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a link
      operationId: delete_link
      description: >
        Delete a stored link
      responses:
        "204":
          description: Link deleted successfully
        "404":
          description: Link not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    scope:
      name: scope
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/DocumentBoxScope"
    folder_id:
      name: folder_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/FolderId"
...
    link_id:
      name: link_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/LinkId"
...
    x-tenant-id:
      name: x-tenant-id
      in: header
      description: ID of the tenant the request is for
      required: true
      schema:
        $ref: "#/components/schemas/TenantId"
    x-tenant-env:
      name: x-tenant-env
      in: header
      description: Environment to operate withins
      required: true
      schema:
        type: string
    x-user-id:
      name: x-user-id
      in: header
      description: ID of the user performing the operation (If performed by a user)
      required: false
      schema:
        $ref: "#/components/schemas/UserId"
    x-user-name:
      name: x-user-id
      in: header
      description: Name of the user performing the operation (If performed by a user)
      required: false
      schema:
        type: string
    x-user-image-id:
      name: x-user-image-id
      in: header
      description: Image ID of the user performing the operation (If performed by a user)
      required: false
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request, response body contains the error message
      content: text/plain
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    UserId:
      type: string
      description: ID of a user
    TenantId:
      type: string
      format: uuid
      description: ID of a tenant
...
    FileId:
      type: string
      format: uuid
      description: ID of a file
    FolderId:
      type: string
      format: uuid
      description: ID of a folder
    LinkId:
      type: string
      format: uuid
      description: ID of a link
    DocumentBoxScope:
      type: string
      description: Unique scope for a document box
      minLength: 1
      example: "user:a71502fe-c025-4a06-ae5d-3131535a8237:uploaded"
...
    FolderResponse:
      type: object
      properties:
        folder:
          $ref: "#/components/schemas/FolderWithExtra"
        children:
          allOf:
            - $ref: "#/components/schemas/ResolvedFolderWithExtra"
            - description: >
                All the files, links, folders within the folder along with the path
                to the folder itself
    FolderWithExtra:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/FolderId"
        name:
          type: string
          description: User provided name for the file
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - nullable: true
              description: >
                ID of the parent folder, this is null when the folder is the root 
                folder of a document box
        created_at:
          type: string
          format: date-time
          description: When the folder was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the folder (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the folder has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - document_box
        - created_at
    FileWithExtra:
      type: object
      description: >
        File with additional information about the most recent edit and user information
      properties:
        id:
          $ref: "#/components/schemas/FileId"
        name:
          type: string
          description: Unique ID of the file
        mime:
          type: string
          description: MIME / content type of the file
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder the file is within
        hash:
          type: string
          description: SHA256 hash of the file contents
        size:
          type: number
          description: File size in bytes
        encrypted:
          type: boolean
          description: Whether the file has been detected as encrypted during processing
        created_at:
          type: string
          format: date-time
          description: When the file was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the file (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the file has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - mime
        - folder_id
        - hash
        - encrypted
        - created_at
    LinkWithExtra:
      type: object
      description: >
        Link with additional information about the most recent edit and user information
      properties:
        id:
          $ref: "#/components/schemas/LinkId"
        name:
          type: string
          description: User provided name of the link
        value:
          type: string
          description: The value of the link
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder this link is within
        created_at:
          type: string
          format: date-time
          description: When the link was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the link (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the link has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - value
        - folder_id
        - created_at
...
    User:
      type: object
      description: >
        Snapshot of the external user who performed the action,
        the details for the user are updated every time they
        perform an action
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/UserId"
            - type: string
              description: External ID of the user
        name:
          type: string
          nullable: true
          description: External name of the user if known
        image_id:
          type: string
          nullable: true
          description: External image ID of the user if known
      required:
        - id
...
    ResolvedFolderWithExtra:
      type: object
      description: Contents of a folder and the resolved path to get to the folder
      properties:
        path:
          type: array
          description: Path to the requested folder
          items:
            $ref: "#/components/schemas/FolderPathSegment"
        folders:
          type: array
          description: Folders within the folder
          items:
            $ref: "#/components/schemas/FolderWithExtra"
        files:
          type: array
          description: Files within the folder
          items:
            $ref: "#/components/schemas/FileWithExtra"
        links:
          type: array
          description: Links within the folder
          items:
            $ref: "#/components/schemas/LinkWithExtra"
    FolderPathSegment:
      type: object
      description: Segment of the path to a folder
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder from the segment part
        name:
          type: string
          description: Name of the folder at the segment
    ErrorResponse:
      type: object
      properties:
        reason:
          type: string
          description: Descriptive message for the cause of the error
        backtrace:
          type: string
          description: >
            Server backtrace of the error, only present in local debugging disabled
            for production releases